Refactor summary card component into one shared
These are functionally identical apart from the edit link, so let’s DRY
things up
