+++ lib/engines/content_block_manager/app/components/content_block_manager/content_block/document/show/embedded_objects/summary_cards_component.html.erb
@@ -1,5 +1,5 @@
 <% embedded_objects.keys.each do |key| %>
-  <%= render ContentBlockManager::ContentBlock::Document::Show::EmbeddedObjects::SummaryCardComponent.new(
+  <%= render ContentBlockManager::Shared::EmbeddedObjects::SummaryCardComponent.new(
     content_block_edition:,
     object_type:,
     object_name: key,
