+++ lib/engines/content_block_manager/test/components/shared/embedded_objects/summary_card_component_test.rb
@@ -1,6 +1,6 @@
 require "test_helper"
 
-class ContentBlockManager::ContentBlockEdition::Details::EmbeddedObjects::ConfirmSummaryCardComponentTest < ViewComponent::TestCase
+class ContentBlockManager::Shared::EmbeddedObjects::SummaryCardComponentTest < ViewComponent::TestCase
   extend Minitest::Spec::DSL
   include ContentBlockManager::Engine.routes.url_helpers
 
@@ -19,10 +19,38 @@ class ContentBlockManager::ContentBlockEdition::Details::EmbeddedObjects::Confir
   let(:content_block_edition) { build_stubbed(:content_block_edition, :email_address, details:) }
 
   it "renders a summary list" do
-    component = ContentBlockManager::ContentBlockEdition::Details::EmbeddedObjects::ConfirmSummaryCardComponent.new(
+    component = ContentBlockManager::Shared::EmbeddedObjects::SummaryCardComponent.new(
       content_block_edition:,
       object_type: "embedded-objects",
       object_name: "my-embedded-object",
+    )
+
+    render_inline component
+
+    assert_selector ".govuk-summary-card__title", text: "Embedded Object details"
+
+    assert_selector ".govuk-summary-list__row", text: /Name/ do
+      assert_selector ".govuk-summary-list__key", text: "Name"
+      assert_selector ".govuk-summary-list__value", text: "My Embedded Object"
+    end
+
+    assert_selector ".govuk-summary-list__row", text: /Field 1/ do
+      assert_selector ".govuk-summary-list__key", text: "Field 1"
+      assert_selector ".govuk-summary-list__value", text: "Value 1"
+    end
+
+    assert_selector ".govuk-summary-list__row", text: /Field 2/ do
+      assert_selector ".govuk-summary-list__key", text: "Field 2"
+      assert_selector ".govuk-summary-list__value", text: "Value 2"
+    end
+  end
+
+  it "renders a summary list with edit link" do
+    component = ContentBlockManager::Shared::EmbeddedObjects::SummaryCardComponent.new(
+      content_block_edition:,
+      object_type: "embedded-objects",
+      object_name: "my-embedded-object",
+      show_edit_action: true,
     )
 
     render_inline component
